Moved files around into separate folders for more organized structure
